+++ CHANGELOG.md
@@ -4,6 +4,11 @@ See DataLab [roadmap page](https://datalab-platform.com/en/contributing/roadmap.
 
 ## DataLab Version 0.19.0 ##
 
+💥 New features and enhancements:
+
+* Region of Interest editor:
+  * Regrouped the graphical actions (new rectangular ROI, new circular ROI, new polygonal ROI) in a single menu "Graphical ROI"
+
 🛠️ Bug fixes:
 
 * Fixed [Issue #144](https://github.com/DataLab-Platform/DataLab/issues/144) - Average profile extraction: `ValueError` when selection rectangle is larger than the image